# exo2/src/main/java/io/swagger/api/DeleteEmpruntApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;


import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import javax.validation.constraints.*;
import javax.validation.Valid;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-11-09T11:43:32.706Z")

@Api(value = "deleteEmprunt", description = "the deleteEmprunt API")
public interface DeleteEmpruntApi {

    @ApiOperation(value = "Supprimer un emprunt", notes = "", response = Void.class, tags={ "Emprunt", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Void.class) })
    
    @RequestMapping(value = "/deleteEmprunt/{idEmprunt}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteEmpruntIdEmpruntDelete(@ApiParam(value = "ID du emprunt à supprimer",required=true ) @PathVariable("idEmprunt") Integer idEmprunt);

}